+++ package.json
@@ -52,6 +52,7 @@
     "react-native-safe-area-context": "4.5.0",
     "react-native-screens": "~3.20.0",
     "react-native-svg": "13.4.0",
+    "react-native-url-polyfill": "^1.3.0",
     "react-native-webview": "11.26.0",
     "react-native-webview-crypto": "^0.0.25",
     "react-qr-code": "^2.0.11",
